Update DTOs to support nullable properties for improved flexibility and data handling

+++ src/MainService.Infrastructure/Services/Services.cs
@@ -8,7 +8,9 @@ namespace MainService.Infrastructure.Services
     {
         public async Task<bool> DeleteAsync(Photo item)
         {
-            var itemToDelete = await uow.PhotoRepository.GetByIdAsync(item.Id);
+            Photo? itemToDelete = await uow.PhotoRepository.GetByIdAsync(item.Id);
+
+            if (itemToDelete == null) return false;
 
             if (!string.IsNullOrEmpty(itemToDelete.PublicId))
             {
@@ -29,7 +31,9 @@ namespace MainService.Infrastructure.Services
     public class MedicalInsuranceCompaniesService(IUnitOfWork uow) : IMedicalInsuranceCompaniesService {
         public async Task<bool> DeleteByIdAsync(int id)
         {
-            MedicalInsuranceCompany itemToDelete = await uow.MedicalInsuranceCompanyRepository.GetByIdAsync(id);
+            MedicalInsuranceCompany? itemToDelete = await uow.MedicalInsuranceCompanyRepository.GetByIdAsync(id);
+
+            if (itemToDelete == null) return false;
 
             uow.MedicalInsuranceCompanyRepository.Delete(itemToDelete);
 
@@ -42,7 +46,9 @@ namespace MainService.Infrastructure.Services
     public class PaymentMethodTypesService(IUnitOfWork uow) : IPaymentMethodTypesService {
         public async Task<bool> DeleteByIdAsync(int id)
         {
-            PaymentMethodType itemToDelete = await uow.PaymentMethodTypeRepository.GetByIdAsync(id);
+            PaymentMethodType? itemToDelete = await uow.PaymentMethodTypeRepository.GetByIdAsync(id);
+
+            if (itemToDelete == null) return false;
 
             uow.PaymentMethodTypeRepository.Delete(itemToDelete);
 
@@ -55,7 +61,9 @@ namespace MainService.Infrastructure.Services
     public class SpecialtiesService(IUnitOfWork uow) : ISpecialtiesService {
         public async Task<bool> DeleteByIdAsync(int id)
         {
-            Specialty itemToDelete = await uow.SpecialtyRepository.GetByIdAsync(id);
+            Specialty? itemToDelete = await uow.SpecialtyRepository.GetByIdAsync(id);
+
+            if (itemToDelete == null) return false;
 
             uow.SpecialtyRepository.Delete(itemToDelete);
 
@@ -68,7 +76,9 @@ namespace MainService.Infrastructure.Services
     public class DiseasesService(IUnitOfWork uow) : IDiseasesService {
         public async Task<bool> DeleteByIdAsync(int id)
         {
-            Disease itemToDelete = await uow.DiseaseRepository.GetByIdAsync(id);
+            Disease? itemToDelete = await uow.DiseaseRepository.GetByIdAsync(id);
+
+            if (itemToDelete == null) return false;
 
             uow.DiseaseRepository.Delete(itemToDelete);
 
@@ -81,7 +91,9 @@ namespace MainService.Infrastructure.Services
     public class SubstancesService(IUnitOfWork uow) : ISubstancesService {
         public async Task<bool> DeleteByIdAsync(int id)
         {
-            Substance itemToDelete = await uow.SubstanceRepository.GetByIdAsync(id);
+            Substance? itemToDelete = await uow.SubstanceRepository.GetByIdAsync(id);
+
+            if (itemToDelete == null) return false;
 
             uow.SubstanceRepository.Delete(itemToDelete);
 
@@ -94,7 +106,9 @@ namespace MainService.Infrastructure.Services
     public class OccupationsService(IUnitOfWork uow) : IOccupationsService {
         public async Task<bool> DeleteByIdAsync(int id)
         {
-            Occupation itemToDelete = await uow.OccupationRepository.GetByIdAsync(id);
+            Occupation? itemToDelete = await uow.OccupationRepository.GetByIdAsync(id);
+
+            if (itemToDelete == null) return false;
 
             uow.OccupationRepository.Delete(itemToDelete);
 
@@ -107,7 +121,9 @@ namespace MainService.Infrastructure.Services
     public class MaritalStatusesService(IUnitOfWork uow) : IMaritalStatusesService {
         public async Task<bool> DeleteByIdAsync(int id)
         {
-            MaritalStatus itemToDelete = await uow.MaritalStatusRepository.GetByIdAsync(id);
+            MaritalStatus? itemToDelete = await uow.MaritalStatusRepository.GetByIdAsync(id);
+
+            if (itemToDelete == null) return false;
 
             uow.MaritalStatusRepository.Delete(itemToDelete);
 
@@ -120,7 +136,9 @@ namespace MainService.Infrastructure.Services
     public class EducationLevelsService(IUnitOfWork uow) : IEducationLevelsService {
         public async Task<bool> DeleteByIdAsync(int id)
         {
-            EducationLevel itemToDelete = await uow.EducationLevelRepository.GetByIdAsync(id);
+            EducationLevel? itemToDelete = await uow.EducationLevelRepository.GetByIdAsync(id);
+
+            if (itemToDelete == null) return false;
 
             uow.EducationLevelRepository.Delete(itemToDelete);
 
@@ -135,7 +153,9 @@ namespace MainService.Infrastructure.Services
     public class ColorBlindnessesService(IUnitOfWork uow) : IColorBlindnessesService {
         public async Task<bool> DeleteByIdAsync(int id)
         {
-            ColorBlindness itemToDelete = await uow.ColorBlindnessRepository.GetByIdAsync(id);
+            ColorBlindness? itemToDelete = await uow.ColorBlindnessRepository.GetByIdAsync(id);
+
+            if (itemToDelete == null) return false;
 
             uow.ColorBlindnessRepository.Delete(itemToDelete);
 
@@ -148,7 +168,9 @@ namespace MainService.Infrastructure.Services
     public class RelativeTypesService(IUnitOfWork uow) : IRelativeTypesService {
         public async Task<bool> DeleteByIdAsync(int id)
         {
-            RelativeType itemToDelete = await uow.RelativeTypeRepository.GetByIdAsync(id);
+            RelativeType? itemToDelete = await uow.RelativeTypeRepository.GetByIdAsync(id);
+
+            if (itemToDelete == null) return false;
 
             uow.RelativeTypeRepository.Delete(itemToDelete);
 
@@ -161,7 +183,9 @@ namespace MainService.Infrastructure.Services
     public class ConsumptionLevelsService(IUnitOfWork uow) : IConsumptionLevelsService {
         public async Task<bool> DeleteByIdAsync(int id)
         {
-            ConsumptionLevel itemToDelete = await uow.ConsumptionLevelRepository.GetByIdAsync(id);
+            ConsumptionLevel? itemToDelete = await uow.ConsumptionLevelRepository.GetByIdAsync(id);
+
+            if (itemToDelete == null) return false;
 
             uow.ConsumptionLevelRepository.Delete(itemToDelete);
